Move login & logout commands to cypress-base
