+++ modules/cypress-base/src/commands/index.ts
@@ -20,4 +20,5 @@
 /// <reference types="cypress" />
 /// <reference path="../../types/index.d.ts" />
 
+import "./authentication";
 import "./dom";
