+++ modules/cypress-base/src/constants/index.ts
@@ -20,3 +20,4 @@
 export * from "./header-dom-constants";
 export * from "./login-page-dom-constants";
 export * from "./server-constants";
+export * from "./side-panel-dom-constants";
